fix: params는 클라이언트 컴포넌트에서 사용 불가
